maintain(Naming): apply best practice
