+++ frontend/src/app/app.component.scss
