+++ frontend/src/app/app.component.ts
@@ -8,14 +8,14 @@ import { filter } from 'rxjs/operators';
 import { AddRecipeDialogComponent } from './components';
 
 @Component({
-  selector: 'tasty-root',
-  standalone: true,
-  imports: [CommonModule, RouterOutlet, RouterLink, MatButtonModule, MatIconModule],
-  templateUrl: './tasty.component.html',
-  styleUrl: './tasty.component.scss',
+  selector:        'tasty-root',
+  standalone:      true,
+  imports:         [CommonModule, RouterOutlet, RouterLink, MatButtonModule, MatIconModule],
+  templateUrl:     './app.component.html',
+  styleUrl:        './app.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TastyComponent {
+export class AppComponent {
   readonly dialog = inject(MatDialog);
 
   public openDialog(): void {
