+++ frontend/src/app/app.config.ts
@@ -3,10 +3,10 @@ import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { provideRouter, withComponentInputBinding } from '@angular/router';
 
-import { routes } from './tasty.routes';
+import { routes } from './app.routes';
 
 
-export const tastyConfig: ApplicationConfig = {
+export const appConfig: ApplicationConfig = {
   providers: [
     provideZoneChangeDetection({ eventCoalescing: true }),
     provideRouter(routes, withComponentInputBinding()),
