+++ frontend/src/app/components/add-recipe-dialog/add-recipe-dialog.component.html
@@ -1,0 +1,18 @@
+<h1 mat-dialog-title>Rezept hinzufügen</h1>
+<div mat-dialog-content>
+    <mat-form-field appearance="outline" class="tasty-add-recipe-dialog-input">
+        <mat-label>URL</mat-label>
+        <input type="url" matInput [formControl]="urlFormControl" placeholder="Füge hier eine Rezept-URL ein...">
+        @if (urlFormControl.hasError('pattern') && !urlFormControl.hasError('required')) {
+        <mat-error>Bitte gib eine gültige URL ein</mat-error>
+        }
+        @if (urlFormControl.hasError('required')) {
+        <mat-error>Gib bitte eine Rezept-URL ein</mat-error>
+        }
+    </mat-form-field>
+</div>
+<div mat-dialog-actions>
+    <button mat-button [mat-dialog-close]="undefined">Abbrechen</button>
+    <button mat-flat-button [disabled]="!urlFormControl.valid"
+        [mat-dialog-close]="urlFormControl.value">Hinzufügen</button>
+</div>
