+++ frontend/src/app/components/recipe-detail/recipe-detail.component.html
