+++ frontend/src/app/components/recipe-list/recipe-list.component.scss
