+++ frontend/src/app/model/recipe.ts
