+++ frontend/src/app/pipes/index.ts
