+++ frontend/src/app/pipes/kebab-case.pipe.ts
